Added a world map with items
The Hex World Map has the interactive items that the player will interact with in the playspace. Any folders called "Low Poly" contain placeholder assets from the Unity asset store and are segregated from the other files so they'll be easier to remove. If someone could add the day/night cycle to it, that'd be awesomesauce.