2 How to play images have been added for the menu
These two images have been added for the menu. The first image, how to play.png is supposed to have text added that talks about the objective and the significance of the timer. And the second one how to play2.png is supposed to talk about how there are robotic enemies trying to kill the player, and that opponents have metallic skin.