Added a logo. Added story pages for main menu.
The "story" should accessible from the main menu. Players click the "story" button, which is a series of images that the players click through via a "next" button. There will be text at the bottom of each image explaining the story. When the player gets to the last image, there should be a button that allows them to return to the main menu.